internal/lsp: add index and channel read snippets
Similar to functions, we now detect if a map, slice, or array's
element type matches the expected type. If so, we give it a snippet
indexing into the object. For maps, we include the map key type as a
placeholder.

I also gave the same treatment to channels. If reading from the
channel yields the expected type, we include a snippet to prepend "<-"
to the channel completion item.