internal/lsp: clear package cache more aggressively
There is a class of bugs where files get "stuck" with errors. I think
this happens mainly when a goFile "forgets" what packages it is
associated with, causing the goFile's DidChange events to fail to
invalidate anything from the package cache.

Address this by always invalidating the package cache in
loadParseTypecheck. This method is only called when there is already a
need to re-parse or re-type check a file's package.

Invalidating the package cache also clears out the goFile's "pkgs"
map, so I removed the redundant code from runGopackages(). In general
I think it is unsafe to clear a goFile's "pkgs" but not clear the
pcache since that can lead to orphaned goFiles.

We also no longer need the special logic to clear the pcache when a
file has been added or deleted to an existing package since we always
clear the pcache now.